Pre-Migration Work * Move ServerConfig to config * Move Constants to service sub-package * Move some constants from PbjBlockStreamServiceProxy to Constants
