+++ server/src/main/java/org/hiero/block/server/config/ServerConfig.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package org.hiero.block.server;
+package org.hiero.block.server.config;
 
 import com.swirlds.config.api.ConfigData;
 import com.swirlds.config.api.ConfigProperty;
