+++ server/src/test/java/org/hiero/block/server/BlockNodeAppTest.java
@@ -16,6 +16,7 @@ import io.helidon.webserver.WebServerConfig;
 import io.helidon.webserver.http.HttpRouting;
 import java.io.IOException;
 import java.util.List;
+import org.hiero.block.server.config.ServerConfig;
 import org.hiero.block.server.config.logging.ConfigurationLogging;
 import org.hiero.block.server.consumer.ConsumerConfig;
 import org.hiero.block.server.events.BlockNodeEventHandler;
